old fix to loss func

+++ loss/segmentation.py
@@ -121,7 +121,7 @@ def pixel_crossentropy_loss(
     """
     assert len(y_pred.shape) > 2, "y_pred must have at least 3 dimensions."
     batch_size, num_classes = y_pred.shape[:2]
-    # y_true = y_true.long()
+    y_true = y_true.long()
 
     if mode == "auto":
         if y_pred.shape == y_true.shape:
